Kubernetes: display "`:100:`" not "💯"
GitHub, when displaying Markdown, has the unfortunate tendency to
display ":100:" as the emoji "💯".

This is a problem because ":100:" appears in our IPv6 addresses. Our fix
is to prepend a "0", i.e. ":0100:".

This is an easier fix than inserting backslashes (e.g. "\:100\:", which
aren't treated consistently across different Markdown renderers (e.g.
GitHub, Atom)--sometimes the backslashes are rendered.