🐞 k8s: fix kube-scheduler `x509: certificate signed by unknown authority`
We somehow skipped the step where we created `kube-scheduler`'s
configuration file. Now we properly create it & point to it via
`/etc/kubernetes/scheduler`'s `KUBE_SCHEDULER_ARGS` environment variable
which passes the `--config` flag.

I don't know how we ever got this to work in the first place because
we never created the config file according to my notes.